Protect special characters in PDF render backend requests
Since we are passing a URL in a query string, we need to double-escape
special characters to ensure the preservation of special characters like
question marks.

This patch uses the existing, but internal _encodeURIComponent function.
This is primarily for expediency. We could also consider promoting
encodeURIComponent to a public function in swagger-router, and then
using that version.

Bug: T169223