Reading lists: add sorting
Add 'sort' parameter to /lists/ and /lists/{id}/entries/,
set sort/dir for MW API accordingly.

Also add sort parameters to MW API requests sent by
/lists/changes/since/{date} (has no effect since those are
the default settings, but good for self-documentation).

(Also get rid of a bunch of old FIXME comments. I wanted to see
if it's possible to just return an object with multiple keys
from a helper method and merge that with the rest of the query
parameters, but I'm getting used to the current syntax.)

Bug: T180402