Lower down rate limit for all the Analytics Edit metrics to 25 rps
We have recently been hit by a bot making 100rps (already rate limited)
scanning the whole mw-history and causing a bit of pressure to the Druid
backend. Druid is not as fast as Cassandra for some type of queries
(especially the ones for big time windows), so it makes sense in my
opinion to lower down the maximumg rps allowed for the moment.