Include `pagelanguage` into the title_revision table.
For language variants support it's not enough to look
whether the domain language supports multiple variants,
since on every wiki each page can actually have it's own
lang, and since LanguageConverter is enabled universally,
the individual pages could be transformed even on wikis
where we generically don't support variants.

Thus, we need to include `pagelanguage` property into
the title_revision table in order to be able to check,
whether to try the transformtion, or just to skip it.
Also will be useful to do Varnish cache splitting more
intelegently.

Bug: https://phabricator.wikimedia.org/T190689
Change-Id: I92052e3f7980d005f1df8810f3641e92bdc9fc9b