Allow http URLs in example config
The x-sub-request-filters patterns used to allow http in the past.
A couple of weeks ago this was change to only allow https. The problem
with this is that developers need to manually change these patterns
when using a local backend instead of the public ones.

This patch bring back the ability to use http://localhost for backend
services, like MCS or Parsoid, without an extra step needed to get the
configuration right.

Bug: https://phabricator.wikimedia.org/T202425