Remove per-editor metrics endpoints
The per-editor metrics have been discussed in some RfCs and
community has agreed that, except for english wikipedia where
metrcsi per-editor are available, on other project editors
should opt-in for their metrics to be visible.
Analytics-Team will discuss with Community-Liaison to see if
the endpoints should be reenabled or fully forgotten.

Bug: T203826