Remove old hack we reusing TID.
During the switch from the old storage to the new storage we've used
to use old storage instead of actually running Parsoid, so we've
needed to reuse the TID of the old content before storing the new
content. The hack was removed, but the reusing code remained. It's not
needed any more without the hack.

Change-Id: Ie4356cd93e83955d37ac697407724da2a6024509