Disable onthisday automatic check as it causes RB deployment failures
This is a temporary measure to make RESTBase deployment succeed. In the
mid-term, we should at least turn this check into a test so as to ensure
the output adheres to the specification. In the long run, we should
investigate how to make the check not fail and resurrect it.

Bug: T203588