Removed accept-language duplicate
Added a varyDoesNotContainDuplicates function in test/utils/assert.js that throws an error whenever there is a duplicate.

Added a test in test/features/pagecontent/language_variants.js that checks to see if there are language duplicates in vary

Changed the addVaryHeader method in lib/mwUtil.js to check if a version of the vary exists before adding a new one.