Consolidate sys/ projects
Until now, we would just be adding paths to sys/ in each project as they
would change. However, most of these paths for most projects would be
exactly the same. This commit consolidates them and groups them in three
sub-projects:

- business: a sub-project containing the business-logic modules
  (parsoid, events, etc.)
- backend-storage: a sub-project containing back-end storage logic
  (table, action, page_revisions, key_value)
- backend-proxy: a sub-project containing sys/ paths that need proxying
  from the front-end service to the back-end service.

This commit also introduces the `service_type` configuration variable
that is used to choose between which modules to load. Its possible
values are `frontend`, `backend` and `fullstack`. For now, it selects
between backend-storage and backend-proxy modules, but in the future it
will also be used to decide whether to load the business sub-project
too.