Add the v1 proxy for key_value and page_revisions
v1/proxy.js adds the `v1/key_value` and `v1/page_revisions` end points
for backend and fullstack configurations and it keeps the specification
unchanged for frontend configurations.

It reuses the sys/backend-proxy sub-project for the end point
definitions, thus ensuring the symmetry between frontend and fullstack
configurations.

The backend_proxy sys module is made generic to allow requests coming
for both v1/ and sys/ namespaces. Additionally, it ensures that if it is
publicly exposed, a 403 is returned for external requests.