[TEMP] Parsoid: Log all VE requests
VE receives weird 404s from RESTBase during html2wt transforms, so log
all VE requests to try to figure out what is going on.

Note: this commit impacts performance as it checks the user agent on
each call to /sys/parsoid, so it will need to be reverted once we
collect enough data.

Bug: T233127