Parsoid: Ensure the returned ETag is the one we stash the data under
We use the HTML's response's ETag to stash the data, but return the
overall response's ETag. While they are equal in most cases, ensure that
the proper ETag is returned without counting on the magic of them being
equal.

Also, in case we don't have an ETag to stash and return, error out as
it's better to fail in that case than create a new one since that means
that either we have incomplete data or that Parsoid did not return a
proper response (or both).

Bug: T234928