Parsoid: Use the ETag's revid to stash the content under
Clients may want to stash the contents of a document regardless of
whether they provide the revid in the URI or not. If they do, then the
contents being stashed needs to match that revid (ensured by
`generateAndSave`), but if they don't then we have to use the revid
present in the ETag. They either have to match or we must use the one in
the ETag as the client has not supplied it. Therefore, always use the
one found in the response's ETag.

Bug: T234928