Parsoid: If a stash can't be found, retry with undefined as the revid
When the client doesn't provide the revid to stash the contents under,
the key used by RESTBase is `{title}:undefined:{tid}`, so retry using
that key before falling back to fetching the latest contents.

Note: this is only a temprary work-around for the current situation in
storage where a considerable amount of stashes are stored with this key.
This should be removed after 24h.

Bug: T234928