Parsoid: Stash: Honour our own ETag when retrieving the stash
If a client is switching between wikitext and HTML and in one of the two
transforms they fail to provide the revision in the URI, trust that the
ETag is correct and try to retrieve the stashed content using the
information in the ETag. As a fall-back, though, try with the
client-supplied revision as well.

This is a port of #1208 to the proxy branch.

Bug: T235465