Parsoid: Proxy: Mirror traffic in split mode as well
Once we enter split mode, we still want to keep mirroring all of the
traffic to both variants as we want the storage for both variants to
contain the latest renders.

Also, if the client explicitly set the variant, then do not fall back to
the other regardless of the mode.