Remove duplicate filter definitions
When hyperswitch loads a route, it automatically assigns it the filters
loaded for its hierarchical parent. This can cause problems in cases
where the filter is not idempotent, as is the case of the language
variant filter which performs an extra action (calling Parsoid for
conversion); we then end up chaining the action twice, each using the
result of the previous one.

Bug: T234266