Parsoid: Proxy: Split fall-back and mirror request cloning
When we are in split mode, fall back to the invert regardless of whether
the variant is the default one or not. This allows us to easily use the
split mode as an invert of itself.

Also, clone the request to be used in the mirroring so that headers
going to the different variants do not contain the added
`X-Parsoid-Variant` header.

Bug: T229015