mobile-html-offline-resources: make 'title' parameter optional
The mobile-html-offline-resources endpoint accepts a required title and
optional revision and tid path parameters despite not using any of them.
This is undesirable from a caching perspective because it will result
in the same content being cached for many different URLs per wiki.

This change makes the title parameter optional so that we can encourage
clients to stop sending title parameters to improve the cache hit rate,
while continuing to support existing clients that send it.

This change depends on a corresponding change to mobileapps itself.
(https://gerrit.wikimedia.org/r/c/mediawiki/services/mobileapps/+/608730)

Bug: https://phabricator.wikimedia.org/T254490