Use disabled_storage setting to disable storage.
Several endpoints already have a way to disable storage. Use the same
machnism for parsoid. There is no need to use a regular expression to
match domains, because the config can be defined for each domain
separately.

Note that the disabled_storage flag is set for some test and beta domains
in production, which would cause parsoid storage to be disabled on these
domains as well when this change is deployed. See https://github.com/search?q=repo%3Awikimedia%2Fmediawiki-services-restbase-deploy%20wikipedia.org_no_storage&type=code

Change-Id: I02fcb262ab303113e2d1282a9d87bffabd6e4cf2